refactor: reorganize directory structure of GUI-related files
