+++ Makefile
@@ -6,7 +6,7 @@ OBJS = \
 	src/kernel/multitask.o \
 	src/kernel/timer.o \
 	src/kernel/datetime.o \
-	src/kernel/graphic.o \
+	src/kernel/GUI/graphic.o \
 	src/kernel/tek.o \
 	src/kernel/utf82kt.o \
 	src/kernel/sysinfo.o \
@@ -16,8 +16,8 @@ OBJS = \
 	src/kernel/HTML/HTMLNode.o \
 	src/kernel/HTML/HTMLTreeBuilder.o \
 	src/kernel/File.o \
-	src/kernel/ModalWindow.o \
-	src/kernel/Tab.o \
+	src/kernel/GUI/ModalWindow.o \
+	src/kernel/GUI/Tab.o \
 	src/kernel/bmp.o \
 	src/kernel/jpeg.o \
 	src/driver/FAT12.o \
