+++ src/kernel/GUI/ModalWindow.cpp
@@ -1,5 +1,5 @@
-#include <cloumo/kernel/graphic.h>
-#include <cloumo/kernel/ModalWindow.h>
+#include <cloumo/kernel/GUI/graphic.h>
+#include <cloumo/kernel/GUI/ModalWindow.h>
 
 ModalWindow::ModalWindow(const string &msg) : sheet(new Sheet(Size(480, 300))), message(msg) {
 	// background and message
